requirements.txt
transformers
numpy
pinecone-client
seaborn
mysql-connector-python
torch
torchvision
torchaudio
seaborn
sqlalchemy
streamlit
pandas
nltk
--extra-index-url https://download.pytorch.org/whl/cu116
